requirements.txt
# Updated Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2

# AI/ML Libraries (Updated)
torch==2.3.1           # Updated from 2.1.1
torchvision==0.18.1    # Updated from 0.16.1
transformers==4.40.1   # Updated from 4.35.2
sentence-transformers==2.7.0  # Updated from 2.2.2
spacy==3.7.4           # Minor update
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
# NLP & Text Processing
nltk==3.8.1
textstat==0.7.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Google AI
google-generativeai==0.3.2

# Graph Neural Networks (optional)
torch-geometric==2.4.0
networkx==3.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Utilities
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.0
python-json-logger==2.0.7
tqdm==4.66.1

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0